+++ lib/ajax-datatables-rails/configuration.rb
@@ -4,7 +4,6 @@ module AjaxDatatablesRails
   class Configuration
     include ActiveSupport::Configurable
 
-    config_accessor(:db_adapter) { :postgresql }
     config_accessor(:nulls_last) { false }
   end
 end