+++ spec/ajax-datatables-rails/orm/active_record_sort_records_spec.rb
@@ -56,7 +56,7 @@ describe AjaxDatatablesRails::ORM::ActiveRecord do
       datatable.params[:order]['0'] = { column: '0', dir: 'asc' }
       datatable.params[:order]['1'] = { column: '1', dir: 'desc' }
       expect(datatable.sort_records(records).to_sql).to include(
-        "ORDER BY users.username ASC #{nulls_last_sql}, users.email DESC #{nulls_last_sql}"
+        "ORDER BY users.username ASC #{nulls_last_sql(datatable)}, users.email DESC #{nulls_last_sql(datatable)}"
       )
     end
   end
@@ -70,7 +70,7 @@ describe AjaxDatatablesRails::ORM::ActiveRecord do
       nulls_last_datatable.params[:order]['0'] = { column: '0', dir: 'asc' }
       nulls_last_datatable.params[:order]['1'] = { column: '1', dir: 'desc' }
       expect(nulls_last_datatable.sort_records(records).to_sql).to include(
-        "ORDER BY users.username ASC, users.email DESC #{nulls_last_sql}"
+        "ORDER BY users.username ASC, users.email DESC #{nulls_last_sql(datatable)}"
       )
     end
   end